pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fitz>=0.0.1.dev2",
    "flask>=3.1.0",
    "kivy>=2.3.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "onnx>=1.17.0",
    "onnxruntime>=1.21.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.0",
    "pymupdf>=1.25.5",
    "pyopengl>=3.1.9",
    "tensorflow-cpu>=2.14.0",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
    "requests>=2.32.3",
]
